Create Ternary-String
You are given a string s such that each its character is either 1, 2, or 3. 
You have to choose the shortest contiguous substring of s such that it contains each of these three characters at least once.
A contiguous substring of string s is a string that can be obtained from s
by removing some (possibly zero) characters from the beginning of s 
and some (possibly zero) characters from the end of s.